Fixed pr comment
